Solution made more atomic.

+++ common/to.etc.alg/src/main/java/to/etc/util/FileLockKeeper.java
@@ -7,8 +7,6 @@ import to.etc.function.SupplierEx;
 
 import java.io.File;
 import java.time.Duration;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeoutException;
 
 /**
@@ -17,39 +15,42 @@ import java.util.concurrent.TimeoutException;
  */
 public class FileLockKeeper {
 
-	private static final Map<String, Long> LOCK_TIMEOUT_MS_MAP = new ConcurrentHashMap<>();
-
 	private static final Logger LOG = LoggerFactory.getLogger(FileLockKeeper.class);
 
 	/**
-	 *  If lock file is present we do waiting route with provided delays and overall timeout.
-	 *  If lock is not present, lock file is created and guarded block is executed.
+	 *  Tries to lock by created a new temp file. If file is already present, we loop in delay periods until file disappears or until specified timeout expires.
 	 */
-	public synchronized static <T> T withLock(String name, Duration delay, @Nullable Duration timeout, SupplierEx<T> retryBlock, SupplierEx<T> guardedBlock) throws Exception {
+	public static <T> T withLock(String name, Duration delay, @Nullable Duration timeout, SupplierEx<T> guardedBlock) throws Exception {
 		File tmpFile = new File(FileTool.getTmpDir(), "FileLockerKeeper_" + name);
-		if(tmpFile.exists()) {
-			if(null != timeout) {
-				Long lockedPeriod = System.currentTimeMillis() - LOCK_TIMEOUT_MS_MAP.get(name);
-				if (lockedPeriod > timeout.toMillis()) {
-					throw new TimeoutException("File lock " + name + " timeout!");
-				}
-			}
-			LOG.warn("File lock " + name + " present at " + tmpFile.getAbsolutePath() + ", delaying " + StringTool.strNanoTime(delay.getNano()));
-			Thread.sleep(delay.toMillis());
-			return retryBlock.get();
-		}
+		Long startedAt = System.currentTimeMillis();
+		boolean lockMade = false;
 		try {
-			if(!tmpFile.createNewFile()) {
-				throw new IllegalStateException("Unable to create lock file??? " + tmpFile.getAbsolutePath());
+			while(!lockMade) {
+				try {
+					lockMade = tmpFile.createNewFile();
+				} catch(Exception ex) {
+					throw new IllegalStateException("Unable to create lock file??? " + tmpFile.getAbsolutePath(), ex);
+				}
+				if(!lockMade) {
+					if(null != timeout) {
+						Long lockedPeriod = System.currentTimeMillis() - startedAt;
+						if(lockedPeriod > timeout.toMillis()) {
+							throw new TimeoutException("File lock " + name + " timeout!");
+						}
+					}
+					LOG.warn("File lock " + name + " present at " + tmpFile.getAbsolutePath() + ", delaying " + StringTool.strNanoTime(delay.getNano()));
+					Thread.sleep(delay.toMillis());
+				}
 			}
 			LOG.warn("Added file lock " + name + ", as " + tmpFile.getAbsolutePath());
 			return guardedBlock.get();
 		}finally {
-			if(!tmpFile.delete()) {
-				throw new IllegalStateException("Unable to delete file lock??? " + tmpFile.getAbsolutePath());
+			if(lockMade) {
+				if(!tmpFile.delete()) {
+					throw new IllegalStateException("Unable to delete file lock??? " + tmpFile.getAbsolutePath());
+				}
+				LOG.warn("Removed file lock " + name + ", as " + tmpFile.getAbsolutePath());
 			}
-			LOCK_TIMEOUT_MS_MAP.remove(name);
-			LOG.warn("Removed file lock " + name + ", as " + tmpFile.getAbsolutePath());
 		}
 	}
 }
